Update Strategy for changes in bolt oauth
This commit updates the strategy to make sure that it works with changes in the bolt oauth

The request phase now directly redirects to the callback_url with the authorization_code, scope and state as the params.
The callback phase then interacts with this data directly than using it from the session.